concept_decomposition/requirements.txt
absl-py==1.4.0
accelerate==0.16.0
aiohttp==3.8.4
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
backcall==0.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
clip @ git+https://github.com/openai/CLIP.git@a9b1bf5920416aaeaec965c25dd9e8f98c864f16
cycler==0.11.0
datasets==2.9.0
debugpy==1.6.6
decorator==5.1.1
diffusers==0.14.0
dill==0.3.6
docker-pycreds==0.4.0
entrypoints==0.4
filelock==3.9.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
ftfy==6.1.1
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.16.1
google-auth-oauthlib==0.4.6
grpcio==1.51.3
huggingface-hub==0.12.1
idna==3.4
imageio==2.26.0
importlib-metadata==6.0.0
ipykernel==6.16.2
ipython==7.34.0
jedi==0.18.2
joblib==1.2.0
jupyter-client==7.4.9
jupyter-core==4.12.0
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.5.3
matplotlib-inline==0.1.6
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
nest-asyncio==1.5.6
nltk==3.8.1
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging==23.0
pandas==1.3.5
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyenchant==3.2.2
Pygments==2.14.0
pyparsing==3.0.9
pyre-extensions==0.0.29
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
responses==0.18.0
rsa==4.9
scikit-learn==1.0.2
scipy==1.7.3
sentry-sdk==1.15.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tokenizers==0.13.2
torch
torchaudio
torchvision
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
transformers==4.26.0
typing-extensions==4.8.0
typing-inspect==0.9.0
urllib3==1.26.14
wandb==0.13.10
wcwidth==0.2.6
Werkzeug==2.2.3
xxhash==3.2.0
yarl==1.8.2
zipp==3.14.0
